% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experiment}
\alias{experiment}
\alias{experiment,simList-method}
\title{Run an experiment using \code{\link[SpaDES.core:spades]{SpaDES.core::spades()}} (moved to SpaDES.project)}
\usage{
experiment(
  sim,
  replicates = 1,
  params,
  modules,
  objects = list(),
  inputs,
  dirPrefix = "simNum",
  substrLength = 3,
  saveExperiment = TRUE,
  experimentFile = "experiment.RData",
  clearSimEnv = FALSE,
  notOlderThan,
  cl,
  ...
)

\S4method{experiment}{simList}(
  sim,
  replicates = 1,
  params,
  modules,
  objects = list(),
  inputs,
  dirPrefix = "simNum",
  substrLength = 3,
  saveExperiment = TRUE,
  experimentFile = "experiment.RData",
  clearSimEnv = FALSE,
  notOlderThan,
  cl,
  ...
)
}
\arguments{
\item{sim}{A \code{simList}.}

\item{replicates}{Number of replicates. See \code{\link[SpaDES.project:experiment]{SpaDES.project::experiment()}}.}

\item{params, modules, objects, inputs}{Alternatives to vary. See
\code{\link[SpaDES.project:experiment]{SpaDES.project::experiment()}}.}

\item{dirPrefix, substrLength, saveExperiment, experimentFile, clearSimEnv}{See
\code{\link[SpaDES.project:experiment]{SpaDES.project::experiment()}}.}

\item{notOlderThan, cl}{Deprecated; see \code{\link[SpaDES.project:experiment]{SpaDES.project::experiment()}}.}

\item{...}{Passed to \code{\link[SpaDES.project:experiment]{SpaDES.project::experiment()}}.}
}
\value{
See \code{\link[SpaDES.project:experiment]{SpaDES.project::experiment()}}.
}
\description{
\strong{Deprecated} -- moved to SpaDES.project.

\code{experiment()} has moved to the \pkg{SpaDES.project} package, which is now
its maintained home. The version here is a thin shim: it emits a deprecation
message and forwards to \code{\link[SpaDES.project:experiment]{SpaDES.project::experiment()}} when that package is
installed. Please update your code to call \code{SpaDES.project::experiment()}
directly.
}
\seealso{
\code{\link[SpaDES.project:experiment]{SpaDES.project::experiment()}}, \code{\link[SpaDES.project:experiment2]{SpaDES.project::experiment2()}}
}
\author{
Eliot McIntire
}
