% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment2.R
\name{experiment2}
\alias{experiment2}
\alias{experiment2,simList-method}
\title{Run experiment, algorithm 2 (moved to SpaDES.project)}
\usage{
experiment2(
  ...,
  replicates = 1,
  clearSimEnv = FALSE,
  createUniquePaths = c("outputPath"),
  useCache = FALSE,
  debug = getOption("spades.debug"),
  drive_auth_account,
  meanStaggerIntervalInSecs
)

\S4method{experiment2}{simList}(
  ...,
  replicates = 1,
  clearSimEnv = FALSE,
  createUniquePaths = c("outputPath"),
  useCache = FALSE,
  debug = getOption("spades.debug"),
  drive_auth_account = NULL,
  meanStaggerIntervalInSecs = 1
)
}
\arguments{
\item{...}{One or more \code{simList} objects (and any named arguments to forward
to \code{spades()}, e.g. \code{events}). See \code{\link[SpaDES.project:experiment2]{SpaDES.project::experiment2()}}.}

\item{replicates, clearSimEnv, createUniquePaths, useCache, debug, drive_auth_account, meanStaggerIntervalInSecs}{See \code{\link[SpaDES.project:experiment2]{SpaDES.project::experiment2()}}.}
}
\value{
See \code{\link[SpaDES.project:experiment2]{SpaDES.project::experiment2()}}.
}
\description{
\strong{Deprecated} -- moved to SpaDES.project.

\code{experiment2()} has moved to the \pkg{SpaDES.project} package, which is now
its maintained home. The version here is a thin shim: it emits a deprecation
message and forwards to \code{\link[SpaDES.project:experiment2]{SpaDES.project::experiment2()}} when that package is
installed. Please update your code to call \code{SpaDES.project::experiment2()}
directly.
}
\seealso{
\code{\link[SpaDES.project:experiment2]{SpaDES.project::experiment2()}}, \code{\link[SpaDES.project:experiment]{SpaDES.project::experiment()}}
}
\author{
Eliot McIntire
}
