% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLists-class.R
\docType{class}
\name{simLists-class}
\alias{simLists-class}
\alias{simLists}
\title{The \code{simLists} class}
\description{
This is a grouping of \code{simList} objects. Normally this class will be
made using \code{experiment2}, but can be made manually if there are
existing \code{simList} objects.
}
\section{Slots}{

\describe{
\item{\code{paths}}{Named list of \code{modulePath}, \code{inputPath},
and \code{outputPath} paths. Partial matching is performed. These
will be prepended to the relative paths of each \code{simList}}

\item{\code{.xData}}{Environment holding the \code{simLists}.}
}}

\section{Accessor Methods}{


None yet defined:
\tabular{ll}{
\code{\link[=simList-accessors-envir]{simList-accessors-envir()}} \tab Simulation environment. \cr
}
}

\author{
Eliot McIntire
}
